added crime details header component and misc updates
