+++ client/components/crime-details-list-item.jsx
@@ -3,8 +3,8 @@ import React from 'react';
 export default function CrimeDetailsListItem(props) {
   return (
     <ul onClick={props.setView} className="list-type-none p-0 roboto-font">
+      <li className="font-weight-bold">{props.detail}</li>
       <li>{props.date} @ {props.address}</li>
-      <li>{props.detail}</li>
     </ul>
   );
 }
